edit models.py

+++ lms/users/models.py
@@ -23,8 +23,8 @@ class Solution(models.Model):
     timestamp = models.DateTimeField(verbose_name='Время загрузки')
     task = models.ForeignKey(to=Tasks, on_delete=models.CASCADE, verbose_name='Задание')
     user = models.ForeignKey(to=User, on_delete=models.CASCADE, verbose_name='Пользователь')
-    time = models.DecimalField(default=0.00, decimal_places=2, max_digits=7, verbose_name='Время')
-    memory = models.PositiveIntegerField(default=0, verbose_name='Память')
+    time = models.FloatField(default=0.00, verbose_name='Время')
+    memory = models.FloatField(default=0.00, verbose_name='Память')
     score = models.DecimalField(default=0.00, decimal_places=2, max_digits=7, verbose_name='Результат')
     status = models.CharField(max_length=150, verbose_name='Статус решения пользователя')
